+++ packages/pangea-sdk/PangeaCyber.Net/Share/Requests/ShareLinkListRequest.cs
@@ -6,6 +6,10 @@ namespace PangeaCyber.Net.Share.Requests
     ///
     public class ShareLinkListRequest : BaseRequest
     {
+        /// <summary>The bucket to use, if not the default.</summary>
+        [JsonProperty("bucket_id")]
+        public string? BucketID { get; set; }
+
         ///
         [JsonProperty("filter")]
         public FilterShareLinkList Filter { get; set; } = new FilterShareLinkList();